GUI issues fixed after a player leaves (using time.sleep)
There are now issues when a player who had a blind exits the game as this shifts the positions of each player in the players lift while the variables which keep track of who has the blinds and buttons aren't adjusted.

I will try to use a different method to determine if the betting round is over instead.